+++ lib/stellar_client/client.rb
@@ -4,12 +4,12 @@ module StellarClient
     include APIClientBase::Client.module(default_opts: :default_opts)
     api_action :send_payment
 
-    attribute :host, String
+    attribute :bridge_host, String
 
     private
 
     def default_opts
-      { host: host }
+      { bridge_host: bridge_host }
     end
 
   end