+++ lib/stellar_client/requests/withdraw_request.rb
@@ -1,0 +1,24 @@
+module StellarClient
+  class WithdrawRequest
+
+    include APIClientBase::Request.module
+
+    BODY_ATTRS = %i[
+      type
+      asset_code
+      dest
+    ]
+    attribute :type, String
+    attribute :asset_code, String
+    attribute :dest, String
+
+    private
+
+    def params
+      BODY_ATTRS.each_with_object({}) do |attr, hash|
+        hash[attr] = send(attr)
+      end
+    end
+
+  end
+end