+++ spec/stellar_app/db/migrate/20180829075340_unique_stellar_base_bridge_callbacks_operation_id.stellar_base.rb
@@ -1,0 +1,12 @@
+# This migration comes from stellar_base (originally 20180816135847)
+class UniqueStellarBaseBridgeCallbacksOperationId < ActiveRecord::Migration[5.1]
+  def up
+    remove_index :stellar_base_bridge_callbacks, :operation_id
+    add_index :stellar_base_bridge_callbacks, :operation_id, unique: true
+  end
+
+  def down
+    remove_index :stellar_base_bridge_callbacks, :operation_id
+    add_index :stellar_base_bridge_callbacks, :operation_id
+  end
+end