+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - `#get_toml` to fetch the toml of `host`
+- `#get_withdraw` to get withdrawal details from an Anchor
 
 ## [0.4.0] - 2017-04-05
 ### Added