+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Changed
+- Append `/withdraw` to host/path for `WithdrawRequest`
+
 ## [0.5.0] - 2018-08-29
 ### Changed
 - Renamed from BridgeClient to StellarClient